Make the logic for finding the next invalid game more clever.
Previously, after the initial search for the first invalid game, the new
game scanner would always scan upwards in increments of 5 when it needed
to find the index of the next invalid game. This could result in taking
awhile to scan all the new waiting lobbies when Steam decided to open
25+ new games at once.

Now the search for the next invalid game always uses the exponential
probing strategy. It also takes into account how many errors the Steam
API is returning, to try to avoid erroneously increasing the number of
games scanned.